Se modifico user.controller. validLogin y routes
