+++ app.js
@@ -8,7 +8,7 @@ require('dotenv').config()
 
 const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
-const authRouter = require('./routes/users');
+const authRouter = require('./routes/auth');
 
 const app = express();
 app.use(cors())
